pyproject.toml
[tool.poetry]
name = "test_tty"
version = "0.1.0"
description = ""
authors = ["Konstantin-Dudersky <Konstantin.Dudersky@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "*"

[tool.poetry.dev-dependencies]
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
